refactor(services): extract RelativeDateTime from CommitCard
Pull the relative-time label logic (Today / Yesterday / X Days Ago /
X Months Ago / X Years Ago) out of CommitCard into a dedicated
RelativeDateTime service so the upcoming quick view - and any future
surface - can reuse it. CommitCard.date now delegates to it.

Relative-time cases move to RelativeDateTimeSpec; CommitCardSpec keeps
the formatting coverage it owns.
